refactored server for readability
