+++ src/private/vector_search.js
@@ -7,6 +7,7 @@ app.use(cors());
 
 // Import the necessary modules
 const fetchEmbedding = require("../../helpers/fetch_openai_embeddings_api");
+const { sendApiChatRequest } = require("../../helpers/retrieval_augmented_generation");
 const { fetchOpenAIEmbeddings } = fetchEmbedding;
 const pgvector = require('pgvector/pg');
 const { client, connectClient } = require('../../helpers/connection.js');
@@ -26,13 +27,24 @@ app.post('/search', async (req, res) => {
     // Extract the input text from the request body
     const inputText = req.body.inputText;
 
-    // Get the vectorized input from OpenAI (or your embedding function)
+    // Get the vectorized input from OpenAI
     const jsonInputFromOpenAi = await fetchOpenAIEmbeddings(inputText);
     const vectorizedInputFromUser = jsonInputFromOpenAi.data[0].embedding;
+    const vdbResponse = await vectorSearch(vectorizedInputFromUser);
 
 
-    const response = await vectorSearch(vectorizedInputFromUser);
+    const headersKeys = Object.keys(vdbResponse[0]).filter((key) => !key.includes('_vector'));
+    const headers = headersKeys.join(' | ');
+    const vdbResults = vdbResponse.map(row => headersKeys.map(key => row[key]).join(' | ')).join('\n');
 
+    const ragMessage = `Skriv en respons som finner det mest korresponderende datasettet fra metadata for spørringen:\nSpørring:${inputText}\nVektor Database Resultater:\n${headers}\n${vdbResults}`;
+    const messages = [
+      { role: "user", content: ragMessage }
+    ];
+    const fullRagResponse = await sendApiChatRequest(messages);
+    console.log(`\n\nThis is the full RAG response: ${fullRagResponse}`);
+
+    
     // Search for the closest vectors in the database, and respond with the results
     res.json(response)
